# docs/swagger.yaml
basePath: /
definitions:
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_handlers.LocationPayloadDoc:
    properties:
      address:
        example: Jl. Raya Bekasi No. 1
        type: string
      city:
        example: Bekasi
        type: string
      geofence_radius:
        example: 100
        type: number
      latitude:
        example: -6.2
        type: number
      longitude:
        example: 106.9
        type: number
      minimum_staffing:
        example: 8
        type: integer
      name:
        example: Warehouse Bekasi
        type: string
      province:
        example: Jawa Barat
        type: string
      type:
        example: warehouse
        type: string
    type: object
  internal_handlers.LocationResponseDoc:
    properties:
      address:
        example: Jl. Raya Bekasi No. 1
        type: string
      city:
        example: Bekasi
        type: string
      geofence_radius:
        example: 100
        type: number
      id:
        example: 1
        type: integer
      latitude:
        example: -6.2
        type: number
      longitude:
        example: 106.9
        type: number
      minimum_staffing:
        example: 8
        type: integer
      name:
        example: Warehouse Bekasi
        type: string
      province:
        example: Jawa Barat
        type: string
      type:
        example: warehouse
        type: string
    type: object
...
  internal_handlers.PayrollResponseDoc:
    properties:
      absent_count:
        example: 0
        type: integer
      base_salary:
        example: 5000000
        type: number
      cold_storage_allowance:
        example: 500000
        type: number
      generated_at:
        example: "2026-05-23T10:00:00Z"
        type: string
      gross_salary:
        example: 5625000
        type: number
      id:
        example: 1
        type: integer
      late_count:
        example: 1
        type: integer
      month:
        example: 5
        type: integer
      net_salary:
        example: 5575000
        type: number
      overtime_pay:
        example: 125000
        type: number
      total_attendance:
        example: 22
        type: integer
      total_deduction:
        example: 50000
        type: number
      user_id:
        example: 10
        type: integer
      year:
        example: 2026
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend HRIS dan operasional frozen-food retail dengan JWT auth, geofenced
    attendance + selfie, leave approval with staffing risk, payroll, PPE/asset compliance,
    dan executive reports.
  title: HRIS Monitoring API
  version: "1.0"
paths:
...
  /locations:
    get:
      description: Get all registered locations including type, geofence, and minimum
        staffing.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.LocationResponseDoc'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List locations
      tags:
      - Locations
    post:
      consumes:
      - application/json
      description: Create a new outlet or warehouse location used for geofencing,
        staffing rules, and location-driven payroll logic.
      parameters:
      - description: Location payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.LocationPayloadDoc'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.LocationResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create location
      tags:
      - Locations
  /locations/{id}:
    delete:
      description: Delete a location record.
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete location
      tags:
      - Locations
    get:
      description: Retrieve one location record including geofence and staffing fields.
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.LocationResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get location by ID
      tags:
      - Locations
    put:
      consumes:
      - application/json
      description: Update an existing location including type, coordinates, geofence
        radius, and minimum staffing.
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: integer
      - description: Location payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.LocationPayloadDoc'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.LocationResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update location
      tags:
      - Locations
...
  /payrolls:
    get:
      description: Calculate current user's payroll for a given month and year, including
        warehouse allowance and overtime.
      parameters:
      - description: Month
        in: query
        name: month
        type: integer
      - description: Year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.PayrollResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Calculate or get payroll
      tags:
      - Payroll
  /payrolls/download:
    get:
      description: Generate and download a PDF payslip for the authenticated user
        and selected month/year.
      parameters:
      - description: Month
        in: query
        name: month
        type: integer
      - description: Year
        in: query
        name: year
        type: integer
      produces:
      - application/pdf
      responses:
        "200":
          description: Payslip PDF
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download payslip PDF
      tags:
      - Payroll
  /payrolls/history:
    get:
      description: Return previously generated payroll records for the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.PayrollResponseDoc'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get payroll history
      tags:
      - Payroll
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"